Hook in to MP4 muxer from RReverser/mpegts
This introduces a dependency on jBinary, and require.js in order to load
jBinary. Ideally, this will eventually be replaced by a vanilla-JS
implementation of an MP4 file builder, for speed and greater
self-containedness.

There is some stuttering around segment boundaries; not sure if that's
due to errors in calculating track durations, or a problem with how
segments are swapped on the page.